requirements.txt
discord.py
requests
html5lib
bs4